Added function to get current branch
