+++ src/commands/create.rs
@@ -1,4 +1,4 @@
-use crate::utils::{ConfigLayout, InitLayout};
+use crate::utils::yaml_layouts::{ConfigLayout, InitLayout};
 use std::fs;
 use std::path::{Path, PathBuf};
